backport #19393 to v1.8.x (#19398)
* backport #19393 to v1.x

* other commits

* namespace fix

* added copying lib_api.cc into pip wheel for building extensions

* fixed setup.py

+++ python/mxnet/gluon/block.py
@@ -999,7 +999,7 @@ class HybridBlock(Block):
                                            'added to the parameter dicts.\n'
                                            'Please check the backend.')
 
-                    param = Parameter(name)
+                    param = Parameter(name, dtype=param_data.dtype)
                     param._load_init(param_data, args[0].context)
                 pair = (False, param)
